Updates prompt packet
