+++ Projects/Server.Tests/Tests/Network/Packets/Outgoing/MessagePackets.cs
@@ -1,3 +1,5 @@
+using Server.Prompts;
+
 namespace Server.Network
 {
     public sealed class MessageLocalized : Packet
@@ -123,6 +125,20 @@ namespace Server.Network
         {
             Stream.Write(serial1);
             Stream.Write(serial2);
+        }
+    }
+
+    public sealed class UnicodePrompt : Packet
+    {
+        public UnicodePrompt(Prompt prompt) : base(0xC2)
+        {
+            EnsureCapacity(21);
+
+            Stream.Write(prompt.Serial); // TODO: Does this value even matter?
+            Stream.Write(prompt.Serial);
+            Stream.Write(0);
+            Stream.Write(0);
+            Stream.Write((short)0);
         }
     }
 }
