--- Projects/Server.Tests/Tests/Network/Packets/Outgoing/UnicodePromptTests.cs
@@ -1,37 +1,0 @@
-using System;
-using System.Buffers;
-using Server.Network;
-using Server.Prompts;
-using Xunit;
-
-namespace Server.Tests.Network
-{
-    internal class TestPrompt : Prompt
-    {
-    }
-
-    public class UnicodePromptTests
-    {
-        [Fact]
-        public void TestUnicodePrompt()
-        {
-            var prompt = new TestPrompt();
-            var data = new UnicodePrompt(prompt).Compile();
-
-            Span<byte> expectedData = stackalloc byte[21];
-            var pos = 0;
-
-            expectedData.Write(ref pos, (byte)0xC2);   // Packet ID
-            expectedData.Write(ref pos, (ushort)0x15); // Length
-            expectedData.Write(ref pos, prompt.Serial);
-            expectedData.Write(ref pos, prompt.Serial);
-
-#if NO_LOCAL_INIT
-            expectedData.Write(ref pos, (ulong)0);
-            expectedData.Write(ref pos, (ushort)0);
-#endif
-
-            AssertThat.Equal(data, expectedData);
-        }
-    }
-}
